Moved InetAddressCallback and AccessControlLoginException to gateway.server.spi.
