+++ service/http.proxy/src/test/java/org/kaazing/gateway/service/http/proxy/HttpProxyHeadersIT.java
@@ -31,6 +31,7 @@ import org.kaazing.gateway.server.test.config.GatewayConfiguration;
 import org.kaazing.gateway.server.test.config.builder.GatewayConfigurationBuilder;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
+import org.kaazing.test.util.MethodExecutionTrace;
 
 import java.net.URI;
 
@@ -39,11 +40,9 @@ import static org.junit.rules.RuleChain.outerRule;
 
 public class HttpProxyHeadersIT {
 
-    @Rule
-    public TestRule timeout = new DisableOnDebug(new Timeout(10, SECONDS));
-
-    private final K3poRule robot = new K3poRule();
-
+    private final TestRule timeout = new DisableOnDebug(new Timeout(10, SECONDS));
+    private final TestRule testExecutionTrace = new MethodExecutionTrace();
+    private final K3poRule k3po = new K3poRule();
     private final GatewayRule gateway = new GatewayRule() {
         {
             // @formatter:off
@@ -62,12 +61,12 @@ public class HttpProxyHeadersIT {
     };
 
     @Rule
-    public TestRule chain = outerRule(robot).around(gateway);
+    public final TestRule chain = outerRule(testExecutionTrace).around(k3po).around(gateway).around(timeout);
 
     @Test
     @Specification("http.proxy.headers.remove.hop.by.hop")
     public void removeHopByHopHeaders() throws Exception {
-        robot.finish();
+        k3po.finish();
     }
 
 }
