+++ service/http.proxy/src/test/java/org/kaazing/gateway/service/http/proxy/HttpProxySecureIT.java
@@ -85,8 +85,7 @@ public class HttpProxySecureIT {
 
     private final K3poRule k3po = new K3poRule();
     private final TestRule testExecutionTrace = new MethodExecutionTrace();
-
-    public TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
+    private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
     private final GatewayRule gateway = new GatewayRule() {{
