+++ service/http.proxy/src/test/java/org/kaazing/gateway/service/http/proxy/HttpProxyStreamingIT.java
@@ -43,10 +43,8 @@ import org.kaazing.test.util.MethodExecutionTrace;
 
 public class HttpProxyStreamingIT {
 
-    public final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
-
-    public final TestRule testExecutionTrace = new MethodExecutionTrace();
-
+    private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
+    private final TestRule testExecutionTrace = new MethodExecutionTrace();
     private final K3poRule k3po = new K3poRule();
 
     private final GatewayRule gateway = new GatewayRule() {
